Fix architecture detection on ppc64le
juce_runtime_arch_detection.cpp currently identifies ppc64le as ppc64,
which causes it to use the directory name `ppc64-linux` for VST 3 plugin
contents. However, VST 3 specifies that `uname -m` should be used as the
first component of the directory name, which on 64-bit little-endian
PowerPC is `ppc64le`.

Currently, this causes problems building VST 3 plugins on this platform,
as the VST 3 SDK expects the module directory to be named
`ppc64le-linux`.

This commit adds an additional endianness check when 64-bit PowerPC is
detected, outputting `ppc64` or `ppc64le` as appropriate.